Extract guildAccess.ts, split routes/shared.ts, fix twitch->discord coupling
- Extracted the ~35-line duplicated guild-fetch/access-level-resolution
  logic shared by requireGuildContext (middleware.ts) and POST
  /guild/select (guild.ts) into src/web/guildAccess.ts
  (fetchLiveGuildsForUser/resolveAccessLevelForGuild).
- Split routes/shared.ts (407 lines, unrelated validation/view/error/upload
  helpers) into validation.ts, viewHelpers.ts, errorHandling.ts, and
  uploadMiddleware.ts by responsibility. shared.ts is kept as a temporary
  barrel re-exporting all four, so its ~49 existing importers are
  unaffected here — a follow-up PR repoints them directly and removes it.
- Fixed twitchBot.ts's direct import from discord/voicePresence.ts, the
  only twitch-side module reaching into discord/ outside the project's
  createRuntimeRegistry DI pattern. Added
  twitch/twitchGuildResolutionRuntime.ts and wired it from index.ts,
  matching every other Twitch runtime registration.
