+++ src/index.ts
@@ -4,6 +4,8 @@ import { startTwitchBot, stopTwitchBot, sayInChannel } from './twitch/twitchBot'
 import { getActiveChannels, getActiveChannelUserIds, setChannelJoinedHook } from './twitch/twitchChannelMembership';
 import { startDiscordBot, stopDiscordBot } from './discord/discordBot';
 import { reloadGuildRegistry } from './discord/guildRegistry';
+import { resolveGuildIdForDiscordId } from './discord/voicePresence';
+import { registerTwitchGuildResolutionRuntime } from './twitch/twitchGuildResolutionRuntime';
 import { startTwitchMonitor, stopTwitchMonitor, getMultiTwitchDataForChannel } from './twitch/monitor/twitchMonitor';
 import { startEventSub, stopEventSub, reloadEventSubSubscriptions } from './twitch/eventsub/twitchEventSub';
 import { startEventSubReconciliation, stopEventSubReconciliation } from './twitch/eventsub/twitchEventSubReconciliation';
@@ -114,6 +116,7 @@ async function main(): Promise<void> {
   registerEventSubTwitchRuntime({ send: sayInChannel });
   registerRewardPricingRuntime({ pushPricingUpdate });
   registerTimerCommandsRuntime({ send: sayInChannel, getLoginUserIds: getActiveChannelUserIds });
+  registerTwitchGuildResolutionRuntime({ resolveGuildIdForDiscordId });
 
   // Load the guild registry before the Discord client connects so the
   // messageCreate gate recognises registered guilds from the first message.
