+++ src/twitch/twitchBot.test.ts
@@ -50,7 +50,7 @@ vi.mock('../db', async () => {
   };
 });
 
-vi.mock('../discord/voicePresence', () => ({
+vi.mock('./twitchGuildResolutionRuntime', () => ({
   resolveGuildIdForDiscordId: vi.fn(),
 }));
 
@@ -108,7 +108,7 @@ import {
 } from './twitchChannelMembership';
 import * as twitchChannelMembership from './twitchChannelMembership';
 import { getTwitchEnabledChannels, getAllTwitchLinkedUsers, findUserByTwitchName } from '../db';
-import { resolveGuildIdForDiscordId } from '../discord/voicePresence';
+import { resolveGuildIdForDiscordId } from './twitchGuildResolutionRuntime';
 import { getUsers } from './twitchApi';
 import { setTwitchChannel } from '../shared/statusStore';
 import { executeCustomCommandForTwitch } from '../commands/customCommandHandler';
